LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-sans-fieldset-display-ruby-expected.txt

rubyrt
rubyrt

FAIL .ruby assert_equals: .ruby display expected "ruby" but got "block"
FAIL .rt assert_equals: .rt display expected "ruby-text" but got "block"
